+++ source/io/SampleSourcePcm.c
@@ -105,9 +105,9 @@ size_t sampleSourcePcmRead(SampleSourcePcmData self, SampleBuffer sampleBuffer) 
 static boolByte readBlockFromPcmFile(void* sampleSourcePtr, SampleBuffer sampleBuffer) {
   SampleSource sampleSource = (SampleSource)sampleSourcePtr;
   SampleSourcePcmData extraData = (SampleSourcePcmData)(sampleSource->extraData);
-  unsigned long originalBlocksize = sampleBuffer->blocksize;
+  size_t originalBlocksize = sampleBuffer->blocksize;
   size_t samplesRead = sampleSourcePcmRead(extraData, sampleBuffer);
-  sampleSource->numSamplesProcessed += samplesRead;
+  sampleSource->numSamplesProcessed += (unsigned long)samplesRead;
   return (originalBlocksize == sampleBuffer->blocksize);
 }
